# documents/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "description": "API para gestionar usuarios en el sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/reservation": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Crear una nueva reserva",
                "operationId": "d7dfd146b1023d48a2f5da4a4c777c52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-07"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "member_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "court_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Reservation created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación o lógica de negocio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/{id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Obtener detalles de una reserva",
                "operationId": "098cb2fdf04e790b62cec4f042b441f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Reservation details"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "court": "1",
                                                "member": "1",
                                                "date": "2024-12-07 00:00:00",
                                                "start_time": "2024-12-07 08:00:00",
                                                "end_time": "2024-12-07 09:00:00"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Reserva no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Actualizar una reserva",
                "operationId": "f511754c1fabd60895d1bbcf47115c54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-07"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "member_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "court_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Reservation updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Eliminar una reserva",
                "operationId": "ae11834cc394554c364bbeb28f6fffec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Reservation deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/day": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Obtener reservas de un día específico",
                "operationId": "630d4a545c2d8041076845266bb61049",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha de las reservas",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-22"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservas del día",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Reservations for the day"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "member": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "court": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Court A"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "sport": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Tennis"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-12-22"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "time",
                                                        "example": "08:00:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "time",
                                                        "example": "09:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fecha inválida o sin reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No reservations found for the given date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}